Extract index context menu to a separate module
