+++ templates/index.html.tt2
@@ -40,9 +40,32 @@
 		import * as Index from "mod/index";
 		import * as IndexTable from "mod/index_datatables";
 		import * as LRR from "mod/common";
+		import * as ContextMenu from "mod/index_contextmenu";
 
 		jQuery(() => {
+			[% IF usingdefpass %]
+			LRR.toast({
+				heading: "[% c.lh('You\'re using the default password and that\'s super baka of you') %]",
+				text: '[% c.lh("Login with password \"kamimamita\" and change that shit on the double. ...Or just disable it! Why not check the configuration options afterwards, while you\'re at it?") %]',
+				icon: 'warning',
+				hideAfter: 25000,
+				closeOnClick: false,
+				draggable: false,
+			});
+			[% END %]
+
+			const catList = [];
+			[% IF categories.size > 0 %]
+			[% FOREACH categories %]
+			catList.push({
+				name: "[% name %]",
+				id: "[% id %]",
+			});
+			[% END %]
+			[% END %]
+
 			Index.initializeAll();
+			ContextMenu.initialize(catList);
 			window.Index = Index;
 			window.IndexTable = IndexTable;
 			window.LRR = LRR;
@@ -193,78 +216,6 @@
 	[% INCLUDE footer %]
 
 	<button id="pseudo-copy-btn" data-clipboard-text="" style="display: none">click me</button>
-
-	<script type="module">
-		import * as Index from "mod/index";
-		import I18N from "i18n";
-
-		// Last few remains of JS using server-provided data
-		jQuery(() => {
-
-			[% IF usingdefpass %]
-			//If the json has the "default password" flag, flash a friendly notification inviting the user to change his password
-			LRR.toast({
-				heading: "[% c.lh('You\'re using the default password and that\'s super baka of you') %]",
-				text: '[% c.lh("Login with password \"kamimamita\" and change that shit on the double. ...Or just disable it! Why not check the configuration options afterwards, while you\'re at it?") %]',
-				icon: 'warning',
-				hideAfter: 25000,
-				closeOnClick: false,
-				draggable: false,
-			});
-			[% END %]
-
-			const catList = [];
-			[% IF categories.size > 0 %]
-			[% FOREACH categories %]
-			// Push a tuple into catList
-			catList.push({
-				name: "[% name %]",
-				id: "[% id %]",
-			});
-			[% END %]
-			[% END %]
-
-			// Initialize context menu
-			$.contextMenu({
-				selector: '.context-menu',
-				build: ($trigger, e) => {
-					const id = $trigger.attr("id");
-					const isTankoubon = id && id.startsWith("TANK_");
-					return {
-						callback: function (key, options) {
-							Index.handleContextMenu(key, $(this).attr("id"));
-						},
-						items: {
-							"read": { name: "[% c.lh('Read') %]", icon: "fas fa-book" },
-							...(!isTankoubon ? { "download": { name: "[% c.lh('Download') %]", icon: "fas fa-save" } } : {}),
-							"copy link": { name: "[% c.lh('Copy link') %]", icon: "fas fa-link" },
-							"msm-toggle-archive": {
-								name: Index.selectedArchives.has(id)
-									? I18N.MSMRemoveFromSelection
-									: I18N.MSMAddToSelection,
-								icon: "fas fa-check-square"
-							},
-							[% IF userlogged %]
-							"sep1": "---------",
-							"edit": { name: "[% c.lh('Edit Metadata') %]", icon: "fas fa-pencil-alt" },
-							"delete": { name: "[% c.lh('Delete') %]", icon: "fas fa-trash-alt" },
-								"rating": {
-								    "name": "[% c.lh('Add Rating') %]",
-								    "icon": "fas fa-star",
-								    "items": LRR.loadContextMenuRatings(id)
-								},
-							"collections": {
-								"name": "[% c.lh('Add to Category') %]",
-								"icon": "fas fa-search-plus",
-								"items": LRR.loadContextMenuCategories(catList, id)
-							}
-							[% END %]
-						}
-					}
-				}
-			});
-		});
-	</script>
 
 	<div id="overlay-shade"> </div>
